Split some server_util functions into runtime_util
