+++ lib/streamlit/web/server/server.py
@@ -51,7 +51,6 @@ from streamlit.web.server.routes import HealthHandler
 from streamlit.web.server.routes import MediaFileHandler
 from streamlit.web.server.routes import MessageCacheHandler
 from streamlit.web.server.routes import StaticFileHandler
-from streamlit.web.server.server_util import get_max_message_size_bytes
 from streamlit.web.server.server_util import make_url_path_regex
 from streamlit.web.server.upload_file_request_handler import (
     UploadFileRequestHandler,
@@ -60,6 +59,7 @@ from streamlit.web.server.upload_file_request_handler import (
 from .browser_websocket_handler import BrowserWebSocketHandler
 from .component_request_handler import ComponentRequestHandler
 from .stats_request_handler import StatsRequestHandler
+from ...runtime_util import get_max_message_size_bytes
 
 LOGGER = get_logger(__name__)
